+++ internal/infrastructure/database/dbgen/query.sql.go
@@ -105,7 +105,7 @@ type CreateRouteParams struct {
 	Description        string      `json:"description"`
 	HighlightedPhotoID *int64      `json:"highlighted_photo_id"`
 	Distance           float64     `json:"distance"`
-	Duration           int32       `json:"duration"`
+	Duration           float64     `json:"duration"`
 	ElevationGain      float64     `json:"elevation_gain"`
 	ElevationLoss      float64     `json:"elevation_loss"`
 	PathGeom           interface{} `json:"path_geom"`
@@ -501,7 +501,7 @@ type UpdateRouteParams struct {
 	Description        string      `json:"description"`
 	HighlightedPhotoID *int64      `json:"highlighted_photo_id"`
 	Distance           float64     `json:"distance"`
-	Duration           int32       `json:"duration"`
+	Duration           float64     `json:"duration"`
 	ElevationGain      float64     `json:"elevation_gain"`
 	ElevationLoss      float64     `json:"elevation_loss"`
 	PathGeom           interface{} `json:"path_geom"`
