+++ internal/presentation/route/response.go
@@ -17,7 +17,7 @@ type RouteResponseModel struct {
 	Description        string   `json:"description"`
 	HighlightedPhotoID *int64   `json:"highlighted_photo_id"`
 	Distance           float64  `json:"distance"`
-	Duration           int32    `json:"duration"`
+	Duration           float64  `json:"duration"`
 	ElevationGain      float64  `json:"elevation_gain"`
 	ElevationLoss      float64  `json:"elevation_loss"`
 	Visibility         int16    `json:"visibility"`
