+++ internal/usecase/route/create_route_usecase.go
@@ -50,7 +50,7 @@ type CreateRouteUseCaseInputDto struct {
 	Description        string
 	HighlightedPhotoID *int64
 	Distance           float64
-	Duration           int32
+	Duration           float64
 	ElevationGain      float64
 	ElevationLoss      float64
 	PathGeom           orb.LineString
@@ -68,7 +68,7 @@ type CreateRouteUseCaseOutputDto struct {
 	Description        string
 	HighlightedPhotoID *int64
 	Distance           float64
-	Duration           int32
+	Duration           float64
 	ElevationGain      float64
 	ElevationLoss      float64
 	PathGeom           orb.LineString
