+++ internal/usecase/route/get_route_usecase.go
@@ -53,7 +53,7 @@ type RouteDetaileDto struct {
 	Description        string
 	HighlightedPhotoID *int64
 	Distance           float64
-	Duration           int32
+	Duration           float64
 	ElevationGain      float64
 	ElevationLoss      float64
 	PathGeom           orb.LineString
@@ -75,7 +75,7 @@ type RouteListItemDto struct {
 	Description        string
 	HighlightedPhotoID *int64
 	Distance           float64
-	Duration           int32
+	Duration           float64
 	ElevationGain      float64
 	ElevationLoss      float64
 	Visibility         int16
